create controller , repo, and interface for verify and forgot password
