+++ app/Interface/MentorInterface.php
@@ -9,6 +9,4 @@ interface MentorInterface
     public function login(Request $request);
     public function register(Request $request);
     public function logout();
-    public function verifyMentor($tokenURL);
-    public function forgotPassword($tokenURL);
 }
