+++ app/Interface/UsersInterface.php
@@ -9,8 +9,4 @@ interface UsersInterface
     public function login(Request $request);
     public function register(Request $request);
     public function logout();
-    public function checkVerify($tokenURL);
-    public function verifyUsers($tokenURL);
-    public function forgotPassword(Request $request);
-    public function changePassword($tokenURL, Request $request);
 }
